% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_wrapper.R
\name{MyKmeans}
\alias{MyKmeans}
\title{K-means clustering for a data matrix X}
\usage{
MyKmeans(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{A n x p matrix of numeric data}

\item{K}{The number of clusters.}

\item{M}{(Optional) A K x p matrix of data used as initial centers. If this argument is used, number of rows in M must equal non-optional argument K}

\item{numIter}{The max number of iterations used in K-means clustering algorithm.}
}
\value{
A vector of length $n$ containing integers between 1 and K of cluster assignments.
}
\description{
K-means clustering for a data matrix X
}
\examples{

##### Example 1
set.seed(0)

n1 = 100
n2 = 100
p = 5

X = rbind(matrix(rnorm(n1*p, mean = -2), nrow = n1, ncol = p),
          matrix(rnorm(n2*p, mean = 5),  nrow = n2, ncol = p))

K = 2

cluster_assignment = MyKmeans(X, K)


#####################
##### Example 2

set.seed(0)

n1 = 50
n2 = 50
n3 = 50
n4 = 50

p = 2

X = rbind(matrix(rnorm(n1*p, mean = -3, sd = .1), nrow = n1, ncol = p),
          matrix(rnorm(n2*p, mean = -1, sd = .1), nrow = n2, ncol = p),
          matrix(rnorm(n3*p, mean =  1, sd = .1), nrow = n3, ncol = p),
          matrix(rnorm(n4*p, mean =  3, sd = .1), nrow = n4, ncol = p))

K = 4

M = X[sample.int(n1+n2+n3+n4, size = K, replace = TRUE), , drop = FALSE]
cluster_assignment = MyKmeans(X, K, M)
 
}
